fastapi_server/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
langchain-community
langchain-huggingface
google-cloud-aiplatform
vertexai
sentence-transformers
faiss-cpu
cachetools
torch 